Refactor FileConfigurationPoller slightly...
